+++ src/main/java/com/wildbitsfoundry/etk4j/signals/laplace/Laplace.java
@@ -5,10 +5,7 @@ import com.wildbitsfoundry.etk4j.control.TransferFunction;
 import com.wildbitsfoundry.etk4j.math.complex.Complex;
 import com.wildbitsfoundry.etk4j.math.functions.UnivariateFunction;
 import com.wildbitsfoundry.etk4j.math.polynomials.RationalFunction;
-import com.wildbitsfoundry.etk4j.signals.filters.ButterWorth;
-import com.wildbitsfoundry.etk4j.signals.filters.Elliptic;
-import com.wildbitsfoundry.etk4j.signals.filters.FilterOrderResults;
-import com.wildbitsfoundry.etk4j.signals.filters.FilterSpecs;
+import com.wildbitsfoundry.etk4j.signals.filters.*;
 import com.wildbitsfoundry.etk4j.util.ComplexArrays;
 
 import java.util.Arrays;
@@ -165,6 +162,29 @@ public class Laplace {
         double[] time = {0, 0.000124391246294398, 0.000248782492588796, 0.000373173738883193, 0.000497564985177591, 0.000621956231471989, 0.000746347477766387, 0.000870738724060784, 0.000995129970355182, 0.00111952121664958, 0.00124391246294398, 0.00136830370923838, 0.00149269495553277, 0.00161708620182717, 0.00174147744812157, 0.00186586869441597, 0.00199025994071036, 0.00211465118700476, 0.00223904243329916, 0.00236343367959356, 0.00248782492588796, 0.00261221617218235, 0.00273660741847675, 0.00286099866477115, 0.00298538991106555, 0.00310978115735994, 0.00323417240365434, 0.00335856364994874, 0.00348295489624314, 0.00360734614253753, 0.00373173738883193, 0.00385612863512633, 0.00398051988142073, 0.00410491112771513, 0.00422930237400952, 0.00435369362030392, 0.00447808486659832, 0.00460247611289272, 0.00472686735918712, 0.00485125860548151, 0.00497564985177591, 0.00510004109807031, 0.00522443234436471, 0.00534882359065910, 0.00547321483695350, 0.00559760608324790, 0.00572199732954230, 0.00584638857583669, 0.00597077982213109, 0.00609517106842549, 0.00621956231471989, 0.00634395356101429, 0.00646834480730868, 0.00659273605360308, 0.00671712729989748, 0.00684151854619188, 0.00696590979248627, 0.00709030103878067, 0.00721469228507507, 0.00733908353136947, 0.00746347477766387, 0.00758786602395826, 0.00771225727025266, 0.00783664851654706, 0.00796103976284146, 0.00808543100913585, 0.00820982225543025, 0.00833421350172465, 0.00845860474801905, 0.00858299599431345, 0.00870738724060784, 0.00883177848690224, 0.00895616973319664, 0.00908056097949104, 0.00920495222578543, 0.00932934347207983, 0.00945373471837423, 0.00957812596466863, 0.00970251721096303, 0.00982690845725742, 0.00995129970355182, 0.0100756909498462, 0.0102000821961406, 0.0103244734424350, 0.0104488646887294, 0.0105732559350238, 0.0106976471813182, 0.0108220384276126, 0.0109464296739070, 0.0110708209202014, 0.0111952121664958, 0.0113196034127902, 0.0114439946590846, 0.0115683859053790, 0.0116927771516734, 0.0118171683979678, 0.0119415596442622, 0.0120659508905566, 0.0121903421368510, 0.0123147333831454, 0.0124391246294398, 0.0125635158757342, 0.0126879071220286, 0.0128122983683230, 0.0129366896146174, 0.0130610808609118, 0.0131854721072062, 0.0133098633535006, 0.0134342545997950, 0.0135586458460894, 0.0136830370923838, 0.0138074283386782, 0.0139318195849725, 0.0140562108312669, 0.0141806020775613, 0.0143049933238557, 0.0144293845701501, 0.0145537758164445, 0.0146781670627389, 0.0148025583090333, 0.0149269495553277, 0.0150513408016221, 0.0151757320479165, 0.0153001232942109, 0.0154245145405053, 0.0155489057867997, 0.0156732970330941, 0.0157976882793885, 0.0159220795256829, 0.0160464707719773, 0.0161708620182717, 0.0162952532645661, 0.0164196445108605, 0.0165440357571549, 0.0166684270034493, 0.0167928182497437, 0.0169172094960381, 0.0170416007423325, 0.0171659919886269};
         double[] yout = el.step(time);
         System.out.println(Arrays.toString(yout));
+
+        // Specs for band pass filter
+        FilterSpecs.BandPassSpecs bpSpecs = new FilterSpecs.BandPassSpecs();
+        // The bandwidth of the filter starts at the LowerPassBandFrequency and
+        // ends at the UpperPassBandFrequency. The filter has lower stop band
+        // which is set LowerStopBandFrequency and the upper stop band can be set
+        // with UpperStopBandFrequency. The attenuation at the stop bands can be
+        // set with the LowerStopBandAttenuation and UpperStopBandAttenuation
+        // respectively. In a frequency spectrum, the order of the frequencies will be:
+        // LowerStopBandFrequency < LowerPassBandFrequency < UpperPassBandFrequency <
+        // UpperStopBandFrequency
+        bpSpecs.setLowerPassBandFrequency(190.0); // 190 Hz lower pass band frequency
+        bpSpecs.setUpperPassBandFrequency(210.0); // 210 Hz upper pass band frequency
+        bpSpecs.setLowerStopBandFrequency(180.0); // 180 Hz lower stop band frequency
+        bpSpecs.setUpperStopBandFrequency(220.0); // 220 Hz upper stop band frequency
+        bpSpecs.setPassBandRipple(0.2); // 0.2 dB gain/ripple refer to note
+        bpSpecs.setStopBandAttenuation(20.0); // 20 dB attenuation in the stop band
+
+        FilterOrderResults.OrderAndCutoffFrequencies nW0W1 = ButterWorth.buttord(bpSpecs);
+
+        nW0W1 = Elliptic.ellipord(bpSpecs);
+        TransferFunction elbp = Elliptic.newBandPass(nW0W1.getOrder(), bpSpecs.getPassBandRipple(),
+                bpSpecs.getStopBandAttenuation(), nW0W1.getLowerCutoffFrequency(), nW0W1.getUpperCutoffFrequency());
 
     }
 }
